agents/daily-news/requirements.txt
# Daily News Agent — Python Dependencies
# Install: pip install -r requirements.txt

python-telegram-bot==20.7
openai==1.3.9
requests==2.31.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
